Store files in B2 cloud storage. (#7)
This introduces the B2 cloud storage as a backend. The storages
can be configured with the environment variable `UNITFM_ENV`.
Credentials for B2 have to be passed with `B2_ID` and `B2_SECRET`.

We also test the integration with the new `make integration` target. A
special B2 account was set up to not conflict with and accidentally
delete production data.